Add extension
